requirements.txt
langchain
openai
streamlit
pandas
yfinance
matplotlib
ta
python-dotenv